// src/types/character/Skill.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Ability } from "./Ability";

export type Skill = { name: string, isProficient: boolean, hasExpertise: boolean, abilityName: Ability, additionalMods: number, totalMod: number, };